refactor: 베스트 독서감상문 부분 별도 컴포넌트 분리 (#3)
