+++ src/components/composites/index.ts
@@ -1,1 +1,2 @@
 export * from './login';
+export * from './main';
