+++ src/components/composites/main/bestCommentary/BestCommentary.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { useRecoilValue } from 'recoil';
+
+import { deviceState } from '@/stores';
+import BestCommentaryMobile from './mobile/BestCommentaryMobile';
+import BestCommentaryDesktop from './desktop/BestCommentaryDesktop';
+
+const BestCommentary = () => {
+  const device = useRecoilValue(deviceState);
+
+  return device === 'mobile' ? (
+    <BestCommentaryMobile />
+  ) : (
+    <BestCommentaryDesktop />
+  );
+};
+
+export default BestCommentary;
