+++ src/components/composites/main/bestCommentary/desktop/BestCommentaryDesktop.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const BestCommentaryDesktop = () => {
+  return <div>BestCommentaryDesktop</div>;
+};
+
+export default BestCommentaryDesktop;
