+++ src/components/composites/main/bestCommentary/index.ts
@@ -1,0 +1,1 @@
+export { default as BestCommentary } from './BestCommentary';
