+++ src/components/composites/main/index.ts
@@ -1,0 +1,1 @@
+export * from './bestCommentary';
